ndvm kernel command line: remove "swiotlb=force" option
This otion forces the usage of DMA "bounce buffers" which is
undesirable for performance and stability reasons.  It also
appears to break tg3 devices (packet data is garbage).  This
may be due to a bug somewhere else, but this option should
not be set regardless.

Replace "swiotlb=force" with "iommu=soft".

OXT-11

Signed-off-by: Chris Patterson <pattersonc@ainfosec.com>