Implement 'usb-auto-passthrough' VM property
Provides implementation for the Get and Set properties for the
'usb-auto-passthrough' property for the interfaces:
  com.citrix.xenclient.xenmgr.vm
  com.citrix.xenclient.xenmgr.vm.unrestricted
The intent is to provide means for preventing the automatic passthrough
of new devices to the current in-focus VM on a per-VM basis. The default
behavior for this is to allow the passthrough, which is the previous
behavior.

OXT-64

Signed-off-by: Kevin Pearson <kevin.pearson.4@us.af.mil>