templates: Update ndvm for partitioned disk image
Specify disk images as entire disks instead of using the Xen's fake
partition method - the real partitions will now be exposed.  With that,
update the kernel-extract lines to pull the kernel out of the first disk
partition.  root= now needs to point to the second disk partition.

While in there, replaces the obsolete xencons option with console.

OXT-1327

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>