xenmgr: Use xl option rtc_timeoffset
time-offset is not a valid XL option - XL thinks it is python.  A
similar XL option is rtc_timeoffset which sets the real time clock
offset in seconds.  I don't know what the old units were for this option
though.  xenvm did not seem to do anything with the value.

OXT-1442

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>