templates: Fix syncvm flask config
The old template passed the flask label directly to xenm through
extra-xenvm "flask-label".  This worked for the xenvm-based system, but
it broke with the XL conversion.  "flask-label" is invalid in the XL
config, so it throws an error about the legacy Python syntax.

Use the proper xenmgr config "flask-label" option so it is passed into
the XL config as seclabel.

OXT-1384

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>