xenmgr: Condition NIC configuration on stubdom && HVM
The current NIC code changes some fields depending on the stubdom
setting.  However, stubdom is only applicable to HVM, so PV and PVH
should always follow the non-stubdom case.

We cannot unilaterally change to only an HVM check because a
stubdom-less HVM does not receive emulated NICs.  Emulated NICs cannot
be networked properly from QEMU in dom0.   A Windows 10 guest with tools
and without a stubdom hangs on the Welcome screen.

OXT-1383

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>