xenmgr: getVmStubdom only return True for HVM
If virt-type is "pvh", but stubdom is True, we hit a hang when
booting a VM.  The last log message is "testing pv addons
un-installation" and the domain is paused.

Since stubdom is True, we enter updateStubDomainId and wait for a
stubdom which will never come.  There is a time out, but it's for 1
hour!  The code says the long timeout is for snapshots, so at least it's
explained.

The stubdom setting only matters for HVMs, so conditionalize
getVmStubdom on that.

OXT-1383

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>