xenmgr: Add xl trigger power to HVM shutdown
The QEMU ACPI hvm-shutdown and hvm-power-button xenstore entries don't
always work.  On the QEMU side, it looks like the IRQ is raised and the
ACPI_SLP_BIT is cleared, but the guest doesn't shutdown.  `xl shutdown
-F -w` just hangs.

Manually running `xl trigger DomU power` after that seems to trigger the
shutdown, so just have xenmgr make that invocation.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>