xenmgr: Re-order xl disk spec with target at the end.
Switch to the newer key-value xl disk spec format.  This make it clearer
which keys are which.  Also, to use this format, target needs to go at
the end since the parser grabs the remainder of the string.  This
reduces the redundency of backendtype= strings and is preparatory for
further changes.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>