template/uivm: do not expose xenfb xenstore nodes
libxl will create the backend nodes to trigger the frontend xenfb driver
in the guest to load and initiate the connection with the backend.
vGlass provides a different PV framebuffer driver (openxtfb) which loads
later in the guest boostrap.

Not exposing a vfb avoids having the frontend waiting to connect to a
non-existent backend.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>