xenmgr: Dbus calls to query/filter gpu devices
Add toolstack support for displayhandler related gpu queries

Provides three new dbus interface calls:
* set-displayhandler-gpu -> takes comma-separated list of pci addrs
* get-displayhandler-gpu -> returns common-separated list of drm card
  Ids
* list-gpus-for-displayhandler -> returns tuple of gpu description, drm
  card Id, and pci addr

Displayhandler should ask for its list of assigned gpus by calling
get-displayhandler-gpus. The UI should call list-gpus-for-displayhandler
for all available gpus for the displayhandler, including current
assignments.
The UI can additionally call get-displayhandler-gpus and compare drm
card IDs to find the description or pci address of the assigned gpu(s).

Gpus designated for passthrough are filtered out of list of available
gpus for displayhandler, and gpus assigned to displayhandler are
filtered out of the list of available gpus for passthrough.

Pci addresses are stored in the db because the drm->pci mapping can
change across boots. Therefore, the translation is done on the fly to
maintain proper assignment.

In the event no gpus are assigned to the displayhandler, we use as
default the first entry in the list of available gpus for the
displayhandler.  Consequently, this gpu is not allowed for passthrough
nor is it returned in the list of available gpus for guest VM
passthrough.

Signed-off-by: Eric Chanudet <chanudete@ainfosec.com>