xenmgr: Avoid generating invalid configs
It would be nice to have ndvm be able to toggle between hvm and pv by
just flipping the virt-type, but that isn't possible today.  xenmgr
needs to be more selective to only write appropriate options into the
xl.cfg for the corresponding virt-type.  This commit attempts to do so.

Centralize and expand the virt-type specific options in virtSpecs.

While Xen supports direct booting HVMs with qemu placing the kernel in
VM memory, OpenXT doesn't have support for direct booting with Linux
stubdoms since we don't have a way to get the kernel to QEMU in the
stubdom to place into the VM.  Don't provide kernel or cmdline for HVMs
- xl throws an error when they are in xl.cfg for an HVM.

"boot" is only relevant for HVM, but PV ignores it.

Note, we replace the use of "extra" with "cmdline" since that is
preferred and matches the option in.

Note 2: Disks should be specified by hda, and they will be exposed as
xvda for PV domains.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>