xenmgr: Set TAPDISK3_CRYPTO_KEYDIR
Set TAPDISK3_CRYPTO_KEYDIR in the environment when launching xl.  It
will remain set through the block script and into tap-ctl where it is
used.

The other approach was to have the block-tap script use xec to call back
into xenmgr.  That could work.  The VM case is easy.  The stubdom case
is more complicated since you'd have to map from the stubdom's domid
back to the VM to get the variables.  Since xenmgr has the info, just
pass it down.  This also takes care of the stubdom case since the same
xl process is exec-ing the hotplug scripts.

Help-from: Chris Rogers <rogersc@ainfosec.com>
Signed-off-by: Jason Andryuk <jandryuk@gmail.com>