xenmgr: Disable HVM default vkb
We have dangling vkbds by default today for NDVM and USBVM since
vkb_device defaults true.  Flip it off when vkbd isn't set.

This effectively requires all user VMs to set vkbd=true so they have the
option of receiving a vkbd.  It's okay to default it on for guest VMs
and the already set uivm.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>