xenmgr\Xl: start for Shutdown or Rebooted
Expand Xl.start to start the domain from state Rebooted in addition to
Shutdown.  This allows the restartVm path, triggered by a
domain-initiated reboot, to re-start the Vm.  Otherwise state Rebooted
is ignored and Xl.start does nothing.

This refactors the bulk of the code into a _start helper to be called
from the two cases.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>