xenmgr: Delete monitorAcpi
monitorAcpi runs continuously per running domain.  It is a thread that
launches xl acpi-state once a second to query the ACPI state.  And due
to an inefficency of Xl.acpiState, it also runs xl domid beforehand.
This adds up.

monitorAcpi is described as running while booting a Vm, but it lasts for
the lifetime of the VM.

strangely, reactVmAcpiUpdate says this:

    This is implemented in xenmgr for several reasons. First, there's
    really only 1 acpi state we care about: s3. Xenvm used to fork a
    thread that polled xen for domain acpi state with xc_hvm_param_get,
    which is BAD since xc_hvm_param_get makes a hypercall.Second, we can
    remove responsibility from inputserver to react to acpi state
    changes (it shouldn't have to deal with that anyway). Third, we
    won't have to patch libxl.

monitorAcpi seems to be a re-implementation of that very thread.  QEMU
is patched by write-acpi-state-to-xenstore.patch to write s3/s0.
reactVmAcpiUpdate now will handle QEMU's xenstore writes for S3 and S0,
so monitorAcpi can be deleted.

monitorAcpi would handle S4 or S5 transitions, but xenmgr isn't acting
on those and QEMU won't write those transitions.  This should be okay.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>