xenmgr: Sythensize VM ACPI state changes
The S3 state transition is handled by the watch on he QEMU-updated
acpi-state xenstore node.  With the removal of the monitorAcpi polling,
we lose the other acpi state transitions.  Since S0 and S5 correspond to
running and shutdown, just synthesize those ACPI state changes from the
corresponding VM states.

S4 is distinct, but it doesn't seem to actually be seen in practice.
That is, a windows VM hibernating just shows up as S5 when shutdown.
There is code to handle S4 and set the VM config's hibernated field to
true, but this does not trigger.

With how monitorAcpi was written, there was no initial S0 state
transition on boot.  The ACPI state change was only see for the
transition to S3/S5.  However, S3->S0 did have the state change.  This
new code does not make that differentiation and will trigger S0 on boot.

This handles calling the run-on-acpi-state-change script, but the DBus
signal may no longer be generated.  There is still one signal, and it
will include the acpi_state, but it seems the old code would have
generated two signals for each of Vm and ACPI state, though there is
redundency there.  Looking at the generated DBus signals, a VM in S3
shows acpi_state 3, but state "running".

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>