xenmgr: Wire-up power-state tracking
QEMU has code to update /local/domain/$domid/power-state when it changes
ACPI S state.

This seems to have been working with xenvm, but no longer works after
the xl conversion.

Expand setupAcpiNode to allow the stubdom permission to write to the
xenstore node.  VmPowerUpdate is added to have xenmgr react to the
power-state changes and update the ACPI state.

A hibernated VM will now show as hibernated in the UI.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>