xenmgr: Re-work rebootVM as shutdown & start
Replace the use of `xl reboot` by `xl shutdown` and `xl start` to
implement the vm reboot DBus command.  xl shutdown needs
on_reboot="destroy" in the xl.cfg file.  This allows us to remove the
xenmgr <-> xl signal synchronization.

Xl.shutdown & RpcAgent shutdowns both wait for the VM to shutdown,
except the `xl shutdown -c` hypercall version which seems to terminate
right quick.  The waitForState call is therefore not really implementing
a timeout out.

The corner case is if the VM has ACPI & no PV drivers, but is configured
not to react to the power button.  In that case we can't do much.  If
the VM then internally reboots, the xenmgr React code will reboot the VM
and the waitForState will eventually timeout.

writeXlConfig is removed as rebootVm -> restartVm -> _startVm
startVmInternal -> bootVm calls writeXlConfig.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>