xenmgr: Re-try mount multiple times
We see consistent failure of a PV NDVM to boot where xenmgr fails to
mount the VHD partition.  /dev/loop0p1 is still labeled device_t, so
mount cannot mount it.

This adds a re-try loop to try again instead of failing.

The addition on udevadm settle was an attempt to address this, but it is
proving insufficient for some reason.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>