xenmgr: Remove config.ui idle-time-threshold property
xenmgr no longer acts on the idle time, xcpmd is supposed to handle
that.  This is no longer used and can be removed.  However, the the UI
still queries this.

Make it just return 0 and drop writes.  Later on the UI can remove the
use and this can be removed here and from the IDL.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>