xenmgr: Make copyFileFromDisk operate read-only
copyFileFromDisk is only used to copy files out of a disk image/VHD, so
it only needs to mount the disk as read-only.  Make that change by
dropping the parameter and hardcoding ro = True.

withMountedDisk is also modified to drop the ro parameter, and is
renamed to withMountedRODisk at the same time to indicate the mounted
disk is only available read-only.  Since it is not used outside
Utility.hs, it's export is dropped.

By tap-ctl create-ing and mounting disks read-only, we pave the way for
calculating the disk measurements over the VHD files and not the tapdev
contents.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>