xenmgr: Change libxl disk-spec for blktap
With libxl supporting blktap/vbd3, we can switch from backendtype=phy to
backendtype=tap.  When doing so, the target= no longer needs an tap:/aio:
prefix, but the format= needs to convey the information.

backendtype=phy is still required for LVM partitions, so a little more
care is needed for specifying it.

Signed-off-by: Jason Andryuk <jandryuk@gmail.com>